do changes

+++ database/migrations/2021_07_23_154326_add_ordering_to_category_ones_table.php
@@ -14,7 +14,7 @@ class AddOrderingToCategoryOnesTable extends Migration
     public function up()
     {
         Schema::table('category_ones', function (Blueprint $table) {
-            $table->integer('ordering')->after('published')->unsigned()->default(1)->index();
+            $table->float('ordering')->after('published')->unsigned()->default(1)->index();
         });
     }
